requirements.txt
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
uvicorn[standard]
opentelemetry-distro
requests
python-dotenv
databases
sqlalchemy
asyncpg
psycopg2